+++ sdk/resourcemanager/consumption/armconsumption/zz_generated_constants.go
@@ -10,7 +10,7 @@ package armconsumption
 
 const (
 	moduleName    = "armconsumption"
-	moduleVersion = "v0.4.0"
+	moduleVersion = "v0.5.0"
 )
 
 // BillingFrequency - The billing frequency.