+++ sdk/resourcemanager/consumption/armconsumption/zz_generated_usagedetails_client.go
@@ -50,8 +50,8 @@ func NewUsageDetailsClient(credential azcore.TokenCredential, options *arm.Clien
 	return client, nil
 }
 
-// List - Lists the usage details for the defined scope. Usage details are available via this API only for May 1, 2014 or
-// later.
+// NewListPager - Lists the usage details for the defined scope. Usage details are available via this API only for May 1,
+// 2014 or later.
 // If the operation fails it returns an *azcore.ResponseError type.
 // scope - The scope associated with usage details operations. This includes '/subscriptions/{subscriptionId}/' for subscription
 // scope, '/providers/Microsoft.Billing/billingAccounts/{billingAccountId}' for
@@ -69,7 +69,7 @@ func NewUsageDetailsClient(credential azcore.TokenCredential, options *arm.Clien
 // for invoiceSection scope, and
 // 'providers/Microsoft.Billing/billingAccounts/{billingAccountId}/customers/{customerId}' specific for partners.
 // options - UsageDetailsClientListOptions contains the optional parameters for the UsageDetailsClient.List method.
-func (client *UsageDetailsClient) List(scope string, options *UsageDetailsClientListOptions) *runtime.Pager[UsageDetailsClientListResponse] {
+func (client *UsageDetailsClient) NewListPager(scope string, options *UsageDetailsClientListOptions) *runtime.Pager[UsageDetailsClientListResponse] {
 	return runtime.NewPager(runtime.PageProcessor[UsageDetailsClientListResponse]{
 		More: func(page UsageDetailsClientListResponse) bool {
 			return page.NextLink != nil && len(*page.NextLink) > 0